refactor: combine shapes
