+++ frontend/src/game/character/states.ts
@@ -2,7 +2,7 @@ import StateConfig from '../entity/typeStateConfig';
 import walk from '../sprites/character/walk';
 import run from '../sprites/character/run';
 import idle from '../sprites/character/idle';
-import Rectangle from '../helperTypes/rectangle';
+import { Rectangle } from '../../shapes/';
 import Box from '../box';
 
 enum CharacterState {
