+++ frontend/src/game/helperTypes/canvasHelper.ts
@@ -1,4 +1,4 @@
-import Point from './point';
+import { Point } from '../../shapes';
 
 type CanvasHelper = {
   c:CanvasRenderingContext2D;
