+++ frontend/src/game/sprites/character/walk/index.ts
@@ -1,6 +1,6 @@
 import sprite from './index.png';
 import SpriteConfig from '../../typeConfig';
-import Point from '../../../helperTypes/point';
+import { Point } from '../../../../shapes/';
 
 const sc:SpriteConfig = {
   link: sprite,
