+++ frontend/src/shapes/index.ts
@@ -1,0 +1,5 @@
+import Point from './point';
+import Line from './line';
+import Rectangle from './rectangle';
+
+export { Point, Line, Rectangle };
