+++ frontend/src/shapes/point.ts
