+++ frontend/src/shapes/rectangle.ts
