+++ frontend/src/views/canvas2D/index.ts
@@ -1,7 +1,7 @@
 import template from './index.html';
 import style from './canvas.scss';
 import { View } from '..';
-import Point from '../../game/helperTypes/point';
+import { Point } from '../../shapes';
 
 class Canvas2D extends View {
   private readonly bg:HTMLDivElement;
